chore: organise into dirs
