+++ src/launcher/inject.js
